Move dataclasses and ENUMs out of wrapper
